+++ src/App.tsx
@@ -1,11 +1,6 @@
 import { useContext } from "react";
 import { BrowserRouter as Router, Navigate, Route, Routes } from "react-router-dom";
-import AdminHome from "./components/admin/AdminHome";
-import AppointmentRequestById from "./components/admin/AppointmentRequestById";
-import AppointmentRequests from "./components/admin/AppointmentRequests";
-import LogIn from "./components/admin/LogIn";
-import RequestList from "./components/admin/RequestList";
-import SetAvailableAppointments from "./components/admin/SetAvailableAppointments";
+import AdminContainer from "./components/admin/AdminContainer";
 import Aftercare from "./components/Aftercare";
 import Header from "./components/Header";
 import LoadingDotsIcon from "./components/loading/LoadingDotsIcon";
@@ -16,27 +11,19 @@ import AppContext from "./context/AppContext";
 
 function App() {
   // CONTEXT
-  let { isLoading } = useContext(AppContext);
+  let { user, isLoading } = useContext(AppContext);
 
   return (
     <div className="App">
       <Router>
         {isLoading && <LoadingDotsIcon />}
-        <Header />
+        {!user && <Header />}
         <Routes>
           <Route path="/" element={<Main />} />
           <Route path="/aftercare" element={<Aftercare />} />
           <Route path="/request-appointment" element={<RequestPage />} />
           <Route path="/request-submitted" element={<RequestSubmitted />} />
-          <Route path="/admin/login" element={<LogIn />} />
-          <Route path="/admin/home" element={<AdminHome />} />
-          <Route path="/admin/appointment-requests/*" element={<AppointmentRequests />}>
-            <Route path="new" element={<RequestList />}></Route>
-            <Route path="new/:id" element={<AppointmentRequestById />} />
-            <Route path="rejected" element={<RequestList />} />
-            <Route path="rejected/:id" element={<AppointmentRequestById />} />
-          </Route>
-          <Route path="/admin/set-available-appointments" element={<SetAvailableAppointments />} />
+          <Route path="/admin/*" element={<AdminContainer />} />
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
       </Router>
