+++ src/components/admin/AdminHome.tsx
@@ -1,12 +1,6 @@
-import { useContext } from "react";
-import AppContext from "../../context/AppContext";
 import AdminPage from "./AdminPage";
-// import "./AdminHome.css";
 
 function AdminHome() {
-  // CONTEXT
-  let { user } = useContext(AppContext);
-
   return (
     <AdminPage title="Home">
       <div className="AdminHome">
