+++ src/context/AdminContext.tsx
@@ -1,0 +1,20 @@
+import { createContext } from "react";
+import { AppointmentRequest } from "../models/AppointmentRequest";
+
+interface AdminContextModel {
+  appointmentRequests: AppointmentRequest[];
+  newAppointmentRequests: AppointmentRequest[];
+  rejectedRequests: AppointmentRequest[];
+  handleAppointmentRequests: () => void;
+}
+
+const defaultValue: AdminContextModel = {
+  appointmentRequests: [],
+  newAppointmentRequests: [],
+  rejectedRequests: [],
+  handleAppointmentRequests: () => {},
+};
+
+const AdminContext = createContext(defaultValue);
+
+export default AdminContext;
